remove client from utils
